Improve code structure: separate constants, fix error handling, clean up entry points
- Extract display constants, icon/font mappings into src/constants.py
- Remove display concerns (CANVAS_SIZE, TMPFS_PATH, SATURATION) from config.py
- Eliminate fragile Config.__new__() dummy — pass error_message to render() instead
- Reuse DisplayController via module-level singleton to avoid repeated GPIO allocation
- Wrap watcher.py main loop in main() with __name__ guard
- Use set_values() batch write in updateConfig.py (was 9 individual writes)
